Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -2,7 +2,7 @@
 #include<string.h>
 struct Day{
     char name[100];
-    float temp;
+    int temp;
 };
 struct week{
     struct Day days[7];
